users

+++ src/main/java/com/example/dwspractica/RatingRESTController.java
@@ -5,6 +5,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.Collection;
 
 @RestController
@@ -15,6 +16,8 @@ public class RatingRESTController {
     GameService gameService;
     @Autowired
     RatingService ratingService;
+    @Autowired
+    UserService userService;
 
     @GetMapping("/getRatings/{idGame}")
     public ResponseEntity<Collection<Rating>> showRatings(@PathVariable int idGame) {
@@ -26,15 +29,20 @@ public class RatingRESTController {
     }
 
     @DeleteMapping("/deleteRating/{idGame}/{id}")
-    public ResponseEntity<Rating> deleteRating(@PathVariable int idGame, @PathVariable int id) {
+    public ResponseEntity<Rating> deleteRating(@PathVariable int idGame, @PathVariable int id, HttpServletRequest request) {
         if (!gameService.containsGame(idGame)) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         } else {
             if (!ratingService.containsRating(idGame, id)) {
                 return new ResponseEntity<>(HttpStatus.NOT_FOUND);
             } else {
-                ratingService.deleteRating(idGame, id);
-                return new ResponseEntity<>(HttpStatus.OK);
+                Rating aux= ratingService.getRatings(idGame, id);
+                if (request.isUserInRole("ADMIN")||aux.getId_ur()== userService.getIdFromName(request.getUserPrincipal().getName())){
+                    ratingService.deleteRating(idGame, id);
+                    return new ResponseEntity<>(HttpStatus.OK);
+                }else{
+                    return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+                }
             }
         }
     }
